+++ src/main/kotlin/step3/racing/RacingProcess.kt
@@ -1,0 +1,9 @@
+package step3.racing
+
+interface RacingProcess {
+    fun execute(
+        carCount: Int,
+        attemptCount: Int,
+        minRandomValue: Int,
+    ): Array<Array<Int>>
+}
