+++ src/main/kotlin/step3/racing/RacingProcessFactory.kt
@@ -1,0 +1,7 @@
+package step3.racing
+
+object RacingProcessFactory {
+    fun newInstance(randomGenerator: RandomGenerator): RacingProcess {
+        return RacingProcessImpl(randomGenerator)
+    }
+}
