+++ src/main/kotlin/step3/racing/RandomGenerator.kt
@@ -1,0 +1,5 @@
+package step3.racing
+
+interface RandomGenerator {
+    fun generator(range: IntRange): Int
+}
