+++ src/main/kotlin/step3/racing/RandomGeneratorFactory.kt
@@ -1,0 +1,7 @@
+package step3.racing
+
+object RandomGeneratorFactory {
+    fun newInstance(): RandomGenerator {
+        return RandomGeneratorImpl()
+    }
+}
