+++ src/main/kotlin/step3/racing/RandomGeneratorImpl.kt
@@ -1,0 +1,7 @@
+package step3.racing
+
+internal class RandomGeneratorImpl : RandomGenerator {
+    override fun generator(range: IntRange): Int {
+        return range.random()
+    }
+}
