+++ src/main/kotlin/step3/racing/controller/RacingControllerFactory.kt
@@ -1,0 +1,19 @@
+package step3.racing.controller
+
+import step3.racing.RacingProcess
+import step3.racing.view.InputView
+import step3.racing.view.ResultView
+
+object RacingControllerFactory {
+    fun newInstance(
+        inputView: InputView,
+        resultView: ResultView,
+        racingProcess: RacingProcess,
+    ): RacingController {
+        return RacingControllerImpl(
+            inputView,
+            resultView,
+            racingProcess,
+        )
+    }
+}
